+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All user visible changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
-## 0.4.0 : [unreleased]
+## 0.4.1 : [unreleased]
+
+### changes
+
+* re-export `DomainEventView` in crate root
+
+### docs
+
+* inline documentation of re-exports within crate root
+
+## 0.4.0 : 2020-05-23
 
 ### breaking changes